CoffeeScript and Compass processed & cleaned by Gradle incremental build
I think this is the way to go – low level Grunt tasks executed by Gradle and properly tracked with inputs & outputs.

This commit is sufficient for running but for distribution we'll need to get the *cssmin*, *concat*, *uglify*, *copy*, *rev* & *usemin* Grunt tasks to run against the contents of `src/ratpack/public` and output to somewhere under `build` which then gets zipped up into the distribution.

Following that there will probably be some cruft in the Gruntfile that can be removed as Gradle will be managing the higher-level tasks.

The only thing I'm not sure about is file-watching. For now I can use CodeKit or run the Grunt task in another terminal but ultimately that's just duplicating project config that Gradle should be responsible for.