a few changes to the submitted getting started guide
replaced some references to I as tomas lin

the install java link was inconsistent compared to the rest of the manual

There was a line break missing before reference projects that made it not appear correctly as a list

Peter Ledbrook has merged the pull request for a ratpack template that works, so removed that line. 

Link to mongo module was flipped.