ratpack-datasource: a module for handling comnnections to multiple `java.sql.DataSource`s.
This module exposes a singleton `org.ratpackframework.datasource.DataSourceProvider` that
knows how to initialize and keep track of multiple datasources. The configuration for each
dataSource is kept in either a `.properties` or `.groovy` file whose name must be `datasource`
exactly. The format is very similar to the one found in Griffon's datasource plugin, for
example

    dataSource {
      url = 'jdbc:h2:mem:default'
      driverClassName = 'org.h2.Driver'
      username = 'sa'
      password = ''

      pool {
        maxWait = 60000
        maxIdle = 5
        maxActive = 8
      }
    }

The configuration file must be found in the classpath. This module may also initialize
the schema if a configuration flag is given `schema = create` and a valid `schema.ddl`
file is found in the classpath.