Add workaround for race condition in tests resulting in `java.net.SocketException: Socket closed` errors
It looks like on some machines (currently, only reproducible on David Carr's laptop), there may be a race
condition in the tests between the application under test binding to the socket, and the test http client
attempting to connect to said socket.  Thus far, we haven't been successful in identifying the root cause
of this race condition to fix it.  For now, I'm introducing a slight pause when running the tests to allow
the tests to pass on this machine.  When we identify the true fix, we'll likely want to remove this pause.