Add EmbeddedApp fromServerFunction(Function<? super RatpackServer.Definition.Builder, ? extends RatpackServer.Definition> definition).
This was initially proposed as fromServer(...), but that would result in ambiguity with fromServer(Factory<? extends RatpackServer> server) (at least in the Groovy test I was working with).
Adding signatures to GroovyEmbeddedApp to handle Closures doesn't seem to fix this case, as the methods for closures would have the same erasure.