View support in JsonRender(er).
Added a viewClass property to JsonRender and implemented support for it
in JsonRenderer.

If the view class is non-null it will be used to create a new
ObjectWriter using that view from the supplied ObjectWriter (either
default or supplied in the JsonRender). If the view class is null
everything will work as before.

This fixes #733