Support hot reloading in the Groovy Console. Closes #778.
The issue stems from the fact that the Groovy Console starts a new
Thread for each invocation of the script in the console. Thus a
ThreadLocal won't contain a reference to the previous launch of the
server. We get around this by making the assumption that if the script
being launched is not backed by a physical file of some kind (as is the
case when running from the console), then we maintain a single JVM wide
static reference to the executed server. In all other cases, we use a
ThreadLocal as we can be fairly certain that the same thread is used to
initiated the original server and any reloads.

The limitiation of this approach is that when running from the Groovy
Console only a single RatpackServer can be launched as each subsequent
launch will shut down the previous server.