Start using netty ssl facilities instead jdk native.
This provides a more standard way to setup ssl, using
almost identical ssl/tls related files (certificate and
private key as nginx or apache).

JDK keystores are not well extended and this approach
is more painless for users that just want serve the
application under ssl without the need to know all the
idiosyncrasy of the jdk keystores.