Create a file system for root of JAR if app is running from archive
If user runs a Ratpack app with "java -jar ..." the root of the
classpath is a jar:file:... URI, and this can be converted to a
java.nio.Path, but only with care. This change creates the underlying
Filesystem if necessary and ensures that the JAR URI is in fact valid
before returning it to the Ratpack config.

Users can then use the JAR to store static assets (e.g. in /static
or /public in a Spring Boot app).

Fixes gh-848