Log cause of failure during decoding HTTP request
It is quite painful to debug reasons why Netty/Ratpack fails to decode
the HTTP request in production/staging environments. It would be handy
to log request's method and path, but this simple log message should
be enough.

Example would be that Ratpack responds with same status code and response
body for exceeding limit of request line and for exceeding limit for length
of headers.