Add support for binary websocket frames
The `WebSocket` now has `sendText(String/ByteBuf)` and
`sendBinary(ByteBuf)` methods. The old `send(String/ByteBuf)` methods
are deprecated to make it clear that `sendText()` or `sendBinary()`
should be used instead.

The `WebSocketMessage` now has new `isBinary()` and `getContent()`
methods. The `getText()` method is still supported and decodes the
content into a utf8 `String`. The only drawback of this approach is
that existing implementations of `WebSocketHandler` will start
receiving binary frames in their `onMessage(WebSocketMessage)` methods.
An alternative would be to introduce a new `onBinaryMessage` method
with an empty default implementation, deprecate `onMessage`, and
delegate to `onMessage` from the default implementation of
`onTextMessage` for backwards compatibility.

Resolves #746.