A more consistent DefaultRequest.getUri()
It helps to see Request-URI per the HTTP spec to understand this
change:

Request-URI    = "*" | absoluteURI | abs_path | authority

Previously abs_paths would be passed through unmodified while
absoluteURIs were parsed and either rejected because of malformed
URIs (triggering HTTP 500s) or accepted but with subtle encoding
issues.

"*" was also not properly handled by getPath(). There were also
other implementation-specific quirks like "uri:foo" yielding a
path consisting of the "null" string literal.

The "authority-only" case doesn't need to be handled because
Ratpack does not support the CONNECT method.

This change leaves abs_paths alone but tries to make all the
other cases more consistent with one another. When in doubt
we try to do something predictable. This does, however,
change the semantics of a few edge cases. I have kept the
commit implementing tests independent of the key change
here so the previous semantics and the new semantics can
be compared.

This is the proposed fix/work-around for #1517.