Fix spin loop with concurrent yield of caching promise
If a downstream yields on a cached promise while a previous downstream is already pushing the value downstream, the later yield will recurse in a loop until all of downstream dispatching is finished.

This was caused by yields without the lock incorrectly checking if they need to check again after missing the lock. Only the yield that actually takes the lock needs to do this after giving it up.

Closes #1533