Allow HTTP proxy settings to be overridden at request-time
Proxy settings influence how the client connects to the destination. Thus, the
proxy settings need to be part of the key used to pool channels (connections).

The result of this is that proxy connections, including their tunnels, are
pooled. Thus, when using proxies, the client does not always have to establish
a new tunnel to the destination.

The proxy settings are only added to the HttpChannelKey if there are proxy
settings. Thus, is no additional memory overhead in the common case where
proxies are not used.