Add support for connecting to HTTP Proxies with HTTPS
Since the proxy connection needs to be done very early on (when the channel is
first created), the SSL context must also be set early on. Thus, the SslContext
is also attached to the HttpChannelKey. However, the SslContext is _not_ used
in HttpChannelKey equality checks. SslContext does not implement equals and
hashcode, and writing a Comparator to do so or similar would be complex.
Instead, Proxy.protocol is the only new property that influences equality.

This means that some request-level changes to the SslContext will _not_ result
in a new connection being established with the proxy. There are some rare use
cases that this might create problems for (such as if SSL client authentication
is in use, and different client certificates result in different proxy
behavior), but these are consistent with Ratpack's HttpCient SSL support in
general (direct connections also have this same problem).