Don't continuously check whether keepalives are needed when there is no demand (#1691)
This triggered what is effectively a busy wait in that we scheduled another immediate check. Now, if there is no demand we do not scheduled another check. When demand comes in, we either send a heartbeat if no items are sent in response to new demand or restart the keep alive check if they were.