Add basic support for HTTP pipelining (#1700)
Pipelined requests are not processed concurrently. Instead, they are queued. This is also important for a potential race condition with streaming responses over keep-alive requests where the client sends the next request before the server pipeline has fully completed the response handling and resetting the connection state to handle the next request.